Move to public trajectory interface with private TOTP implementation
