+++ include/common_robotics_utilities/time_optimal_trajectory_parametrization.hpp
@@ -1,214 +1,70 @@
 #pragma once
 
-/* ORIGINAL LICENSE TEXT:
- *
- * Copyright (c) 2011, Georgia Tech Research Corporation
- * All rights reserved.
- *
- * Author: Tobias Kunz <tobias@gatech.edu>
- * Date: 05/2012
- *
- * Humanoid Robotics Lab      Georgia Institute of Technology
- * Director: Mike Stilman     http://www.golems.org
- *
- * Algorithm details and publications:
- * http://www.golems.org/node/1570
- *
- * This file is provided under the following "BSD-style" License:
- *   Redistribution and use in source and binary forms, with or
- *   without modification, are permitted provided that the following
- *   conditions are met:
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *   * Redistributions in binary form must reproduce the above
- *     copyright notice, this list of conditions and the following
- *     disclaimer in the documentation and/or other materials provided
- *     with the distribution.
- *   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
- *   CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
- *   INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- *   MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- *   DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- *   CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- *   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- *   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF
- *   USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- *   AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- *   LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- *   ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- *   POSSIBILITY OF SUCH DAMAGE.
- */
-
-#include <stdlib.h>
-#include <iostream>
-#include <functional>
-#include <Eigen/Geometry>
+#include <memory>
+#include <stdexcept>
 #include <vector>
-#include <chrono>
-#include <list>
+
+#include <Eigen/Geometry>
 
 namespace common_robotics_utilities
 {
 namespace time_optimal_trajectory_parametrization
 {
-class PathSegment
+/// Stores a position+velocity+time point in a trajectory.
+class PositionVelocityTimePoint
 {
 public:
+  PositionVelocityTimePoint() {}
 
-  PathSegment(const double length = 0.0) : length_(length) {}
-
-  virtual ~PathSegment() {}
-
-  double Length() const { return length_; }
-
-  virtual Eigen::VectorXd GetConfig(const double s) const = 0;
-
-  virtual Eigen::VectorXd GetTangent(const double s) const = 0;
-
-  virtual Eigen::VectorXd GetCurvature(const double s) const = 0;
-
-  virtual std::list<double> GetSwitchingPoints() const = 0;
-
-  virtual PathSegment* Clone() const = 0;
-
-  double position;
-
-protected:
-
-  double length_;
-};
-
-
-
-class Path
-{
-public:
-
-  Path(const std::list<Eigen::VectorXd>& path,
-       const double maxDeviation = 0.0);
-
-  Path(const Path& path);
-
-  ~Path();
-
-  double Length() const;
-
-  Eigen::VectorXd GetConfig(const double s) const;
-
-  Eigen::VectorXd GetTangent(const double s) const;
-
-  Eigen::VectorXd GetCurvature(const double s) const;
-
-  double GetNextSwitchingPoint(const double s, bool& discontinuity) const;
-
-  std::list<std::pair<double, bool>> SwitchingPoints() const;
-
-private:
-
-  PathSegment* GetPathSegment(double& s) const;
-
-  double length_;
-
-  std::list<std::pair<double, bool>> switching_points_;
-
-  std::list<PathSegment*> path_segments_;
-};
-
-class Trajectory
-{
-public:
-  // Generates a time-optimal trajectory
-  Trajectory(const std::list<Eigen::VectorXd>& waypoints,
-             const Eigen::VectorXd& max_velocity,
-             const Eigen::VectorXd& max_acceleration,
-             const double max_deviation = 0.0,
-             const double timestep = 0.001);
-
-  Trajectory(const Path& path,
-             const Eigen::VectorXd& max_velocity,
-             const Eigen::VectorXd& max_acceleration,
-             const double timestep = 0.001);
-
-  // Returns the optimal duration of the trajectory
-  double Duration() const;
-
-  // Return the position vector and velocity vector of the
-  // robot for a given point in time within the trajectory.
-  // !!! NOT THREAD SAFE - MUTABLE CACHE INSIDE !!!
-  std::pair<Eigen::VectorXd, Eigen::VectorXd> GetPositionVelocity(
-      const double time) const;
-
-  // Outputs the phase trajectory and the velocity limit curve
-  // in 2 files for debugging purposes.
-  void OutputPhasePlaneTrajectory() const;
-
-private:
-
-  struct TrajectoryStep
+  PositionVelocityTimePoint(
+      const Eigen::VectorXd& position, const Eigen::VectorXd& velocity,
+      const double time)
+      : position_(position), velocity_(velocity), time_(time)
   {
-    TrajectoryStep() {}
+    if (position_.size() != velocity_.size())
+    {
+      throw std::invalid_argument("position_.size() != velocity_.size()");
+    }
+    if (time_ < 0.0)
+    {
+      throw std::invalid_argument("time must be >= 0");
+    }
+  }
 
-    TrajectoryStep(double path_pos, double path_vel)
-      : path_pos_(path_pos), path_vel_(path_vel) {}
+  const Eigen::VectorXd& Position() const { return position_; }
 
-    double path_pos_;
-    double path_vel_;
-    double time_;
-  };
+  const Eigen::VectorXd& Velocity() const { return velocity_; }
 
-  bool GetNextSwitchingPoint(const double path_pos,
-                             TrajectoryStep& next_switching_point,
-                             double& before_acceleration,
-                             double& after_acceleration);
+  double Time() const { return time_; }
 
-  bool GetNextAccelerationSwitchingPoint(const double path_pos,
-                                         TrajectoryStep& next_switching_point,
-                                         double& before_acceleration,
-                                         double& after_acceleration);
-
-  bool GetNextVelocitySwitchingPoint(const double path_pos,
-                                     TrajectoryStep& next_switching_point,
-                                     double& before_acceleration,
-                                     double& after_acceleration);
-
-  bool IntegrateForward(std::list<TrajectoryStep>& trajectory,
-                        const double acceleration);
-
-  void IntegrateBackward(std::list<TrajectoryStep>& start_trajectory,
-                         const double path_pos,
-                         const double path_vel,
-                         const double acceleration);
-
-  double GetMinMaxPathAcceleration(const double path_position,
-                                   const double path_velocity,
-                                   const bool max);
-
-  double GetMinMaxPhaseSlope(const double path_position,
-                             const double path_velocity,
-                             const bool max);
-
-  double GetAccelerationMaxPathVelocity(const double path_pos) const;
-
-  double GetVelocityMaxPathVelocity(const double path_pos) const;
-
-  double GetAccelerationMaxPathVelocityDeriv(const double path_pos);
-
-  double GetVelocityMaxPathVelocityDeriv(const double path_pos);
-
-  std::list<TrajectoryStep>::const_iterator GetTrajectorySegment(
-      const double time) const;
-
-  Path path_;
-  Eigen::VectorXd max_velocity_;
-  Eigen::VectorXd max_acceleration_;
-  uint32_t n_ = 0u;
-  std::list<TrajectoryStep> trajectory_;
-
-  static const double eps;
-  const double time_step_;
-
-  mutable double cached_time_;
-  mutable std::list<TrajectoryStep>::const_iterator cached_trajectory_segment_;
+private:
+  Eigen::VectorXd position_;
+  Eigen::VectorXd velocity_;
+  double time_ = 0.0;
 };
+
+/// Basic interface to a position+velocity trajectory.
+class PositionVelocityTrajectoryInterface
+{
+public:
+  virtual ~PositionVelocityTrajectoryInterface() {}
+
+  virtual std::unique_ptr<PositionVelocityTrajectoryInterface>
+  Clone() const = 0;
+
+  virtual double Duration() const = 0;
+
+  virtual PositionVelocityTimePoint GetPositionVelocityTimePoint(
+      const double time) const = 0;
+};
+
+/// Parametrize the provided @param path with the provided velocity and
+/// acceleration limits into a trajectory using Time-Optimal Trajectory
+/// Parametrization.
+std::unique_ptr<PositionVelocityTrajectoryInterface> ParametrizePathTOTP(
+    const std::vector<Eigen::VectorXd>& path,
+    const Eigen::VectorXd& max_velocity,
+    const Eigen::VectorXd& max_acceleration,
+    const double max_deviation = 0.0, const double timestep = 0.001);
 }  // namespace time_optimal_trajectory_parametrization
 }  // namespace common_robotics_utilities
